+++ roms_tools/tests/test_setup/test_initial_conditions.py
@@ -157,6 +157,7 @@ def test_initial_conditions_creation_with_duplicates(use_dask: bool) -> None:
         grid=grid,
         ini_time=datetime(2012, 1, 1),
         source={"path": [fname1, fname2], "name": "GLORYS"},
+        allow_flex_time=True,
         use_dask=use_dask,
     )
 
@@ -164,6 +165,7 @@ def test_initial_conditions_creation_with_duplicates(use_dask: bool) -> None:
         grid=grid,
         ini_time=datetime(2012, 1, 1),
         source={"path": [fname1, fname1, fname2], "name": "GLORYS"},
+        allow_flex_time=True,
         use_dask=use_dask,
     )